Update Account Balance Datatype to Double

+++ src/models/Account.js
@@ -7,7 +7,7 @@ const Account = sequelize.define('Account', {
         primaryKey: true,
     },
     name: { type: DataTypes.STRING, allowNull: false },
-    balance: { type: DataTypes.FLOAT, allowNull: false, defaultValue: 0.0 },
+    balance: { type: DataTypes.DOUBLE, allowNull: false, defaultValue: 0.0 },
     note: { type: DataTypes.TEXT, defaultValue: '' },
     share: { type: DataTypes.INTEGER, defaultValue: 0 },
     price: { type: DataTypes.FLOAT, defaultValue: 0.0 },
